fix(component): 컴포넌트 분리
